refactore: reorganize core.db package
